Move all messaging files into src/extension/messages
