--- packages/sanes-chrome-extension/src/extension/utils/messages.ts
@@ -1,2 +1,0 @@
-// WTC means "web to content"
-export const WTC_MSG_HELLO = 'wtc_HELLO';
